+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Chex.Color` with a `flip/1`.
 - `Piece.King` is now castling aware and provides the castling square(s) in it's
     `available_moves/3`.
+- `Board.get_piece_name/2` to get just the name of a piece at a square.
 
 ### Changed